updated code

+++ apl-core/src/main/java/com/apollocurrency/aplwallet/apl/core/dao/state/derived/BasicDbTable.java
@@ -129,9 +129,9 @@ public abstract class BasicDbTable<T extends DerivedEntity> extends DerivedDbTab
 
 
     @Override
-    public void trim(int height) {
+    public void trim(int height/*, boolean isSharding*/) {
         if (multiversion) {
-            doMultiversionTrim(height);
+            doMultiversionTrim(height/*, isSharding*/);
         } else {
             super.trim(height);
         }
@@ -179,7 +179,7 @@ public abstract class BasicDbTable<T extends DerivedEntity> extends DerivedDbTab
      *                                                               }</pre>
      *               </p>
      */
-    private void doMultiversionTrim(final int height) {
+    private void doMultiversionTrim(final int height/*, boolean isSharding*/) {
         log.trace("doMultiversionTrim(), height={}", height);
         TransactionalDataSource dataSource = databaseManager.getDataSource();
         if (!dataSource.isInTransaction()) {
@@ -197,28 +197,31 @@ public abstract class BasicDbTable<T extends DerivedEntity> extends DerivedDbTab
                      con.prepareStatement("DELETE FROM " + table + " WHERE db_id = ?");
                  PreparedStatement selectDbIdStatement =
                      con.prepareStatement("SELECT db_id, height " + getDeletedColumnIfSupported() + " FROM " + table + " " + keyFactory.getPKClause())) {
-                log.trace("Select {} time: {}", table, System.currentTimeMillis() - startSelectTime);
-                startDeleteTime = System.currentTimeMillis();
+                log.trace("Select 1. {} time: {} ms", table, System.currentTimeMillis() - startSelectTime);
 
+                Set<Long> keysToDelete = new HashSet<>();
                 while (rs.next()) {
-                    Set<Long> keysToDelete = selectDbIds(selectDbIdStatement, rs);
-                    // TODO migrate to PreparedStatement.addBatch for another db
+                    keysToDelete.addAll( selectDbIds(selectDbIdStatement, rs) );
+                }
+                log.trace("Select 2. {} time: {} ms", table, System.currentTimeMillis() - startSelectTime);
+                // TODO migrate to PreparedStatement.addBatch for another db
+                startDeleteTime = System.currentTimeMillis();
+                if (keysToDelete.size() > 0) {
                     for (Long id : keysToDelete) {
                         deleted += deleteByDbId(pstmtDeleteById, id);
-//                        deleted++;
-                        deleteStm++;
                         if (deleted % 100 == 0) {
                             dataSource.commit(false);
                         }
                     }
+//                    log.debug("Delete for table {} took {} ms", table, System.currentTimeMillis() - startDeleteTime);
                 }
                 dataSource.commit(false);
-                log.trace("Delete time {} for table '{}': deleteStm=[{}], deleted=[{}]", System.currentTimeMillis() - startDeleteTime, table,
-                    deleteStm, deleted);
+                log.trace("Delete table '{}' in {} ms: deleted=[{}]",
+                    table, System.currentTimeMillis() - startDeleteTime, deleted);
             }
             long trimTime = System.currentTimeMillis() - startTime;
-            if (trimTime > 1000) {
-                log.debug("Trim for table {} took {} ms", table, trimTime);
+            if (trimTime > 10) {
+                log.debug("Trim for table {} time {} ms", table, trimTime);
             }
         } catch (SQLException e) {
             throw new RuntimeException(e.toString(), e);
@@ -236,7 +239,7 @@ public abstract class BasicDbTable<T extends DerivedEntity> extends DerivedDbTab
     private Set<Long> selectDbIds(PreparedStatement selectDbIdStatement, ResultSet rs) throws SQLException {
         DbKey dbKey = keyFactory.newKey(rs);
         dbKey.setPK(selectDbIdStatement);
-        Set<Long> keys = new HashSet<>();
+        Set<Long> outputKeys = new HashSet<>();
         int maxHeight = rs.getInt("max_height");
         boolean lastDeleted = false;
         Set<Integer> deleteHeights = new HashSet<>();
@@ -253,20 +256,19 @@ public abstract class BasicDbTable<T extends DerivedEntity> extends DerivedDbTab
                     if (entryDeleted) {
                         deleteHeights.add(currentHeight);
                     }
-                    keys.add(dbId);
+                    outputKeys.add(dbId);
                 }
             }
         }
         // last existing record should be 'deleted' and paired with previously deleted records
         if (deleteHeights.size() % 2 != 0 && lastDeleted) {
-            keys.addAll(lastDbIds);
+            outputKeys.addAll(lastDbIds);
         }
-        return keys;
+        return outputKeys;
     }
 
     private int deleteByDbId(PreparedStatement pstmtDeleteByDbId, long dbId) throws SQLException {
         pstmtDeleteByDbId.setLong(1, dbId);
         return pstmtDeleteByDbId.executeUpdate();
-
     }
 }
